Move App.js and App.css to components folder
