+++ src/App.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import store from './app/store';
-import App from './App';
+import App from './components/App/App';
 
 test('renders learn react link', () => {
   const { getByText } = render(
